Little clean up and hardening of the SignalR code
